Update config.plist
Disabled Thunderbolt support. Though it does work well, it requires a warm boot for drivers to load and the card doesn't drop to USB mode when TB isn't working.

The dock I have swaps to USB operation when not connected to a TB port, so I can use it fine with the Mac Pro and swap it onto the MacBook Pro with no issues when using a normal USB 3.0 card, so that's what I'm doing for now.